% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_comp_rate.R
\name{calculate_comp_rate}
\alias{calculate_comp_rate}
\title{Calculate the comparison rate (jämförelseränta)}
\usage{
calculate_comp_rate(date)
}
\arguments{
\item{date}{date}
}
\value{
comparison rate (double)
}
\description{
Calculate the comparison rate (jämförelseränta)
}
\examples{
calculate_comp_rate(Sys.Date() + 89:91)


}
