% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_json}
\alias{export_json}
\title{Export comparison rates to json file}
\usage{
export_json(offsets, file)
}
\arguments{
\item{offsets}{vector of offsets (days from today)}

\item{file}{output file}
}
\value{
Nothing, used for side effects
}
\description{
Export comparison rates to json file
}
