% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_rates.R
\name{get_ref_rates}
\alias{get_ref_rates}
\title{Get reference rates from konsumenternas}
\usage{
get_ref_rates(trade_date = yesterday())
}
\arguments{
\item{trade_date}{date of trade, default = today}
}
\value{
tuyi
}
\description{
Get reference rates from konsumenternas
}
\examples{
get_ref_rates()


}
