% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gov_bond_level.R
\name{gov_bond_level}
\alias{gov_bond_level}
\title{Title}
\usage{
gov_bond_level(date, ...)
}
\arguments{
\item{date}{Date for which to calculate the government bond level}

\item{...}{passed on to get_ref_rates()}
}
\description{
Title
}
\examples{
gov_bond_level(Sys.Date() + 90)


}
